Create a Firebase Configuration YAML File

This function generates a config.yaml file with Firebase project settings and also updates the project's .Renviron file with the variables.

Usage

create_config(
  api_key,
  projectId,
  authDomain,
  storageBucket,
  databaseURL = NULL,
  appId = NULL,
  output_dir = "."
)

Arguments

api_key

The Firebase API Key (required).

projectId

The Firebase Project ID (required).

authDomain

The Firebase Authentication Domain (required).

storageBucket

The Firebase Storage Bucket (required).

databaseURL

The Firebase Database URL (optional).

appId

The Firebase App ID (optional).

output_dir

A string specifying the directory where the config.yaml file will be saved. Defaults to the current working directory. If the directory does not exist, it will be created.

Value

A message indicating the successful creation of the file and environment variables.

Examples

if (FALSE) { # \dontrun{
create_config(
  api_key = "example-key",
  projectId = "example-id",
  authDomain = "example-auth",
  storageBucket = "example-bucket",
  output_dir = "."
)
} # }